Make the SelectRenderer work for sub-queries
